Incorporate PR feedback.
