+++ src/format.rs
@@ -391,8 +391,7 @@ pub struct WildcardEntry {
     pub user_id: CratesUserId,
     pub start: Spanned<chrono::NaiveDate>,
     pub end: Spanned<chrono::NaiveDate>,
-    #[serde(rename = "suggest-renewal")]
-    pub suggest_renewal: Option<bool>,
+    pub renew: Option<bool>,
     pub notes: Option<String>,
     #[serde(rename = "aggregated-from")]
     #[serde(skip_serializing_if = "Vec::is_empty")]
@@ -418,10 +417,10 @@ impl WildcardEntry {
 
     /// Whether a renewal should be suggested for the entry.
     ///
-    /// If the entry expires before `date` (and `suggest_renewal` isn't `false`) a renewal will be
+    /// If the entry expires before `date` (and `renew` isn't `false`) a renewal will be
     /// suggested.
     pub fn should_suggest_renewal(&self, date: chrono::NaiveDate) -> bool {
-        self.suggest_renewal.unwrap_or(true) && self.end < date
+        self.renew.unwrap_or(true) && self.end < date
     }
 }
