+++ tests/test-project/supply-chain/audits.toml
@@ -14,7 +14,7 @@ criteria = "safe-to-deploy"
 user-id = 3618 # David Tolnay (dtolnay)
 start = "2022-01-01"
 end = "2022-07-01"
-suggest-renewal = false
+renew = false
 
 [[audits.atty]]
 criteria = "safe-to-run"
